requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cryptography==41.0.3
defusedxml==0.7.1
Django==4.0.10
django-allauth==0.54.0
django-basic-auth-ip-whitelist==0.5
django-cors-headers==4.2.0
django-jazzmin==2.6.0
django-mongoengine==0.5.4
django-storages==1.13.2
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.1
idna==3.4
mongoadmin==0.2
mongodbforms==0.3
mongoengine==0.27.0
mysql-connector-python==8.1.0
mysqlclient==2.2.0
oauthlib==3.2.2
protobuf==4.21.12
pycparser==2.21
PyJWT==2.8.0
pymongo==3.12.3
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
sendgrid==6.10.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.2.4
starkbank-ecdsa==2.2.0
typing_extensions==4.7.1
urllib3==2.0.4
